pyproject.toml
[tool.poetry]
name = "code-guru-ai"
version = "0.1.0"
description = ""
authors = ["maksym_korniev <maksymkorniev88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
httpx = "^0.28.0"
python-dotenv = "^1.0.1"
setuptools = "^75.6.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
